+++ code/src/main/java/uk/co/conoregan/showrenamer/util/api/TheMovieDatabase/TheMovieDBConverter.java
@@ -15,14 +15,15 @@
  * along with ShowRenamer.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package uk.co.conoregan.showrenamer.utils.api.converters;
+package uk.co.conoregan.showrenamer.util.api.TheMovieDatabase;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
+import uk.co.conoregan.showrenamer.model.api.ResultContainer;
 import uk.co.conoregan.showrenamer.model.show.Episode;
 import uk.co.conoregan.showrenamer.model.show.Movie;
 import uk.co.conoregan.showrenamer.model.show.Season;
-import uk.co.conoregan.showrenamer.utils.api.ResultContainer;
+import uk.co.conoregan.showrenamer.util.api.APIToShowConverter;
 
 import java.util.ArrayList;
